+++ examples/prisma-subscriptions/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pothos-examples/prisma-subscriptions
 
+## 3.0.0-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @pothos/plugin-prisma@4.0.0-next.1
+
 ## 3.0.0-next.0
 
 ### Major Changes